requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
chromadb
langchain-openai>=0.2.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
langsmith>=0.2.0
pandas
streamlit
python-dotenv
pytest
